work with strings in one file
