+++ Client/inc/Chat.h
@@ -5,7 +5,7 @@
 //10.11.7.8 ansh
 //10.11.7.7 anch
 #define SERVERADDR "10.11.7.7"
-#define SERVERPORT 8555
+#define SERVERPORT 8111
 
 #include <stdio.h>
 #include <unistd.h>
