+++ Server/src/main.c
@@ -4,10 +4,7 @@ int main(int argc, char *argv[]) {
     int socket_desc , client_sock , c , *new_sock;
 	struct sockaddr_in server , client;
 	
-	sqlite3* db;
-	db_open("database/uchat.db", &db);
-	char *pass = strdup(db_get_user_password("bubuk", db));
-	printf("\nbubuk password: %s\n", pass);
+	db_open("database/uchat.db", &db);//open sqlite
 	
 	//Create socket
 	socket_desc = socket(AF_INET , SOCK_STREAM , 0);
@@ -20,7 +17,7 @@ int main(int argc, char *argv[]) {
 	//Prepare the sockaddr_in structure
 	server.sin_family = AF_INET;
 	server.sin_addr.s_addr = INADDR_ANY;
-	server.sin_port = htons( 8555 );
+	server.sin_port = htons(SERVERPORT);
 	
 	//Bind
 	if( bind(socket_desc,(struct sockaddr *)&server , sizeof(server)) < 0)
@@ -46,13 +43,13 @@ int main(int argc, char *argv[]) {
 	while( (client_sock = accept(socket_desc, (struct sockaddr *)&client, (socklen_t*)&c)) )
 	{
 		puts("Connection accepted");
-		write(client_sock, "Do you have account?\n If yes please input 'yes' else input 'no'\n", strlen("Do you have account?\n If yes please input 'yes' else input 'no'\n",))
+		/*write(client_sock, "Do you have account?\n If yes please input 'yes' else input 'no'\n", strlen("Do you have account?\n If yes please input 'yes' else input 'no'\n",))
 		if((read_size = recv(sock_from , client_message , 2000 , 0)) > 0){
 			if(strcmp(client_message, 'yes')){
 				mx_registration(client_sock);
 			}
-		}
-		
+		}*/
+
 		pthread_t sniffer_thread;
 		int *new_sock = malloc(1);
 		*new_sock = client_sock;
@@ -63,7 +60,7 @@ int main(int argc, char *argv[]) {
 		
 		
 		//Now join the thread , so that we dont terminate before the thread
-		pthread_join( sniffer_thread , NULL);
+		//pthread_join( sniffer_thread , NULL);
 		puts("Handler assigned");
 		
 	}
