--- Server/src/mx_strnew.c
@@ -1,16 +1,0 @@
-#include "Chat.h"
-
-char *mx_strnew(const int size) {
-
-    char *str = (char*) malloc((size + 1) * sizeof(char));
-
-    if (str == NULL || size < 0) {
-        return NULL;
-    }
-
-    for (int i = 0; i < size; i++) {
-        str[i] = '\0';
-    }
-    str[size] = '\0';
-    return str;
-}
